# public/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "A basic products API",
      "version": "1.0.0",
      "title": "myRetail API",
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "myretailapi.herokuapp.com/api",
    "basePath": "/v1",
    "schemes": [
      "https",
      "http"
    ],
    "paths": {
      "/products": {
        "get": {
          "summary": "List all products",
          "description": "Array of product objects that contain current pricing information",
          "operationId": "listProducts",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Product"
              }
            }
          }
        }
      },
      "/products/{product_id}": {
        "get": {
          "summary": "Get product by product_id",
          "description": "Returns a single product with current pricing info",
          "operationId": "getProductById",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "product_id",
              "in": "path",
              "description": "TCIN product_id of product to return",
              "required": true,
              "type": "integer",
              "format": "int32"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Product"
              }
            },
            "400": {
              "description": "Invalid product_id supplied"
            },
            "404": {
              "description": "Product not found"
            }
          }
        },
        "put": {
            "summary": "Update product pricing info",
            "description": "Updates the current pricing info for a specified product",
            "operationId": "updateProduct",
            "consumes": [
              "application/json"
            ],
            "produces": [
              "application/json"
            ],
            "parameters": [
              {
                "in": "body",
                "name": "body",
                "description": "Product object that needs to be updated",
                "required": true,
                "schema": {
                  "$ref": "#/definitions/Product"
                }
              }
            ],
            "responses": {
              "400": {
                "description": "Invalid product_id supplied"
              },
              "404": {
                "description": "Product not found"
              },
              "405": {
                "description": "Validation exception"
              }
            }
        }
      },
      "/scrape/movies/": {
        "get": {
          "summary": "Scrape movie data from a Target.com",
          "description": "Enter a Target movie PDP to scrape data from and store in the database",
          "operationId": "scrapeProducts",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "description": "The page query param should be a URL for a movie PDP on Target.com. For example: https://www.target.com/p/empire-of-the-sun-dvd-video/-/A-11700719",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation"
            }
          }
        }
      }
    },
    "definitions": {
      "Product": {
        "type": "object",
        "required": [
          "product_id"
        ],
        "properties": {
          "product_id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "example": "The Big Lebowski (Blu-ray) (Widescreen)"
          },
          "current_price": {
            "type": "object",
            "properties": {
                "currency_code": {
                    "type": "string",
                    "example": "USD"
                },
                "value": {
                    "type": "Decimal128",
                    "example": "9.22"
                }
            }
          }
        }
      }
    }
  }